Add log behavior
